mac_bridge/requirements.txt
fastapi
uvicorn[standard]
pyautogui
Pillow
pyobjc-core
pyobjc-framework-Quartz
pyobjc-framework-Cocoa
pyobjc-framework-ApplicationServices
